# Python history improvements