--- a/Common/Python/PandasConverter.cs
+++ b/Common/Python/PandasConverter.cs
@@ -62,14 +62,7 @@ namespace QuantConnect.Python
 
             foreach (var slice in data)
             {
-                if (dataType == null)
-                {
-                    AddSliceDataToDict(slice, sliceDataDict, ref maxLevels);
-                }
-                else
-                {
-                    AddSliceDataTypeDataToDict(slice, dataType, sliceDataDict, ref maxLevels);
-                }
+                AddSliceDataTypeDataToDict(slice, dataType, sliceDataDict, ref maxLevels);
             }
 
             using (Py.GIL())
@@ -240,42 +233,15 @@ namespace QuantConnect.Python
         }
 
         /// <summary>
-        /// Adds each slice data to the pandas data dictionary
-        /// </summary>
-        private void AddSliceDataToDict(Slice slice, IDictionary<Symbol, PandasData> sliceDataDict, ref int maxLevels)
-        {
-            foreach (var key in slice.Keys)
-            {
-                var baseData = slice[key];
-                var value = GetPandasDataValue(sliceDataDict, key, baseData, ref maxLevels);
-
-                if (value.IsCustomData)
-                {
-                    value.Add(baseData);
-                }
-                else
-                {
-                    var ticks = slice.Ticks.ContainsKey(key) ? slice.Ticks[key] : null;
-                    var tradeBars = slice.Bars.ContainsKey(key) ? slice.Bars[key] : null;
-                    var quoteBars = slice.QuoteBars.ContainsKey(key) ? slice.QuoteBars[key] : null;
-                    value.Add(ticks, tradeBars, quoteBars);
-                }
-            }
-        }
-
-        /// <summary>
         /// Adds each slice data corresponding to the requested data type to the pandas data dictionary
         /// </summary>
         private void AddSliceDataTypeDataToDict(Slice slice, Type dataType, IDictionary<Symbol, PandasData> sliceDataDict, ref int maxLevels)
         {
-            var isTick = dataType == typeof(Tick) || dataType == typeof(OpenInterest);
-            // Access ticks directly since slice.Get(typeof(Tick)) and slice.Get(typeof(OpenInterest)) will return only the last tick
-            var sliceData = isTick ? slice.Ticks : slice.Get(dataType);
-
-            foreach (var key in sliceData.Keys)
+            var isTick = dataType == null || dataType == typeof(Tick) || dataType == typeof(OpenInterest);
+            HashSet<Symbol> _addedBars = new();
+            foreach (var baseData in slice.AllData)
             {
-                var baseData = sliceData[key];
-                PandasData value = GetPandasDataValue(sliceDataDict, key, baseData, ref maxLevels);
+                var value = GetPandasDataValue(sliceDataDict, baseData.Symbol, baseData, ref maxLevels);
 
                 if (value.IsCustomData)
                 {
@@ -283,10 +249,34 @@ namespace QuantConnect.Python
                 }
                 else
                 {
-                    var ticks = isTick ? baseData : null;
-                    var tradeBars = dataType == typeof(TradeBar) ? baseData : null;
-                    var quoteBars = dataType == typeof(QuoteBar) ? baseData : null;
-                    value.Add(ticks, tradeBars, quoteBars);
+                    var tick = isTick ? baseData as Tick : null;
+                    TradeBar tradeBar = null;
+                    QuoteBar quoteBar = null;
+                    if(tick == null)
+                    {
+                        tradeBar = dataType == null || dataType == typeof(TradeBar) ? baseData as TradeBar : null;
+                        if (tradeBar != null)
+                        {
+                            if(!_addedBars.Add(tradeBar.Symbol))
+                            {
+                                continue;
+                            }
+                            slice.QuoteBars.TryGetValue(tradeBar.Symbol, out quoteBar);
+                        }
+                        else
+                        {
+                            quoteBar = dataType == null || dataType == typeof(QuoteBar) ? baseData as QuoteBar : null;
+                            if (quoteBar != null)
+                            {
+                                if (!_addedBars.Add(quoteBar.Symbol))
+                                {
+                                    continue;
+                                }
+                                slice.Bars.TryGetValue(quoteBar.Symbol, out tradeBar);
+                            }
+                        }
+                    }
+                    value.Add(tick, tradeBar, quoteBar);
                 }
             }
         }