… Tests/Python/PandasConverterTests.cs
… Tests/Research/QuantBookHistoryTests.cs