<!--- Provide a general summary of your changes in the Title above -->

#### Description
<!--- Describe your changes in detail -->
- Pandas data frame history response will include all ticks. Updating
  existing tests
- Python pandas converter performance improvement
   - Avoid catching the Py.GIL if not required
   - Avoid reiterating over all the data to filter series
   - `ToPandasDataFrame` avoid creating unnecessary objects for the series index
   - Reduce object type checking

TODO: `Serie.Times` for each series we are adding the same timestamps in 99% of the cases, we could have a share list to reduce the garbage

**Backtesting test single symbol**
```python
    def Initialize(self):
        self.SetStartDate(2013,10, 11)
        self.SetEndDate(2013,10,11)
        symbol = self.AddEquity("SPY", Resolution.Tick).Symbol

        start = time.time()
        t = self.History(Tick, symbol, datetime(2013, 10, 9), datetime(2013,10,11), Resolution.Tick)
        end = time.time()
        self.Quit(f"DataFrame size: {t.size}. Took {end - start}")
```

**PR**
- size: 50892793. Took 54.41 @ 7800 MB
- size: 50892793. Took 54.40 @ 7700 MB
- size: 50892793. Took 53.78 @ 7700 MB


**Master**
- size: 50892793. Took 60.14 @ 7200 MB
- size: 50892793. Took 61.05 @ 7500 MB
- size: 50892793. Took 59.67 @ 7700 MB

**Research test Many Symbols**
To big (too many symbols) so adding as a file
[researchTest.txt](https://github.com/QuantConnect/Lean/files/11606772/researchTest.txt)

**MASTER**
- 1m 33s 4.9 GB
- 1m 34s 5.2 GB
- 1m 36s 5.2 GB

**PR**
- 1m 17s 5.2 GB
- 1m 15s 5.5 GB
- 1m 15s 5 GB

#### Related Issue
<!--- This project only accepts pull requests related to open issues -->
<!--- If suggesting a new feature or change, please discuss it in an issue first -->
<!--- If fixing a bug, there should be an issue describing it with steps to reproduce -->
<!--- Please link to the issue here: -->
Closes https://github.com/QuantConnect/Lean/issues/7199
#### Motivation and Context
<!--- Why is this change required? What problem does it solve? -->
Improve python history
#### Requires Documentation Change
<!--- Please indicate if these changes will require updates to documentation, and if so, specify what changes are required -->
N/A
#### How Has This Been Tested?
<!--- Please describe in detail how you tested your changes. -->
<!--- Include details of your testing environment, and the tests you ran to -->
<!--- see how your change affects other areas of the code, etc. -->
Existing tests and performance tests
#### Types of changes
<!--- What types of changes does your code introduce? Put an `x` in all the boxes that apply: -->
- [x] Bug fix (non-breaking change which fixes an issue)
- [x] Refactor (non-breaking change which improves implementation)
- [x] Performance (non-breaking change which improves performance. Please add associated performance test and results)
- [ ] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing functionality to change)
- [ ] Non-functional change (xml comments/documentation/etc)

#### Checklist:
<!--- The following is a checklist of items that MUST be completed before a PR is accepted -->
<!--- If you're unsure about any of these, don't hesitate to ask. We're here to help! -->
- [x] My code follows the code style of this project.
- [x] I have read the **CONTRIBUTING** [document](https://github.com/QuantConnect/Lean/blob/master/CONTRIBUTING.md).
- [ ] I have added tests to cover my changes. <!--- If not applicable, please explain why -->
- [x] All new and existing tests passed.
- [x] My branch follows the naming convention `bug-<issue#>-<description>` or `feature-<issue#>-<description>`

<!--- Template inspired by https://www.talater.com/open-source-templates/#/page/99 -->